requirements.txt
streamlit==1.42.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.34
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
openai==1.61.1
chromadb==0.6.3
tiktoken==0.8.0
python-dotenv==1.0.1
pdfplumber==0.11.5
pandas==2.2.3
langchain-chroma==0.2.2
pypdf==5.3.0
pydantic==2.10.6
tabulate==0.9.0